Refacto: avoid having huge files, separate pkgs
