+++ sorter/comments.go
@@ -1,0 +1,38 @@
+package sorter
+
+import (
+	"fmt"
+	"go/ast"
+	"go/token"
+	"strings"
+)
+
+type commentsMetadata struct {
+	Doc     *ast.CommentGroup
+	Comment *ast.CommentGroup
+}
+
+type importPosition struct {
+	Start token.Pos
+	End   token.Pos
+}
+
+func (p *importPosition) IsInRange(comment *ast.CommentGroup) bool {
+	if p.Start <= comment.Pos() && comment.Pos() <= p.End {
+		return true
+	}
+
+	return false
+}
+
+func importWithComment(imprt string, commentsMetadata map[string]*commentsMetadata) string {
+	var comment string
+	commentGroup, ok := commentsMetadata[imprt]
+	if ok {
+		if commentGroup != nil && commentGroup.Comment != nil && len(commentGroup.Comment.List) > 0 {
+			comment = fmt.Sprintf("// %s", strings.ReplaceAll(commentGroup.Comment.Text(), "\n", ""))
+		}
+	}
+
+	return fmt.Sprintf("%s %s", imprt, comment)
+}
